Move visitor to its own module
